# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface KpiItem {
  metric: string;
  value: string;
  period: string;
  source: string;
}

export interface DashboardSummary {
  totalPrivateInvestment2024B: number;
  averageAdoptionRate2025: number;
  leadingIndustry: string;
  leadingIndustryRate2025: number;
  kpis: KpiItem[];
  lastUpdated: string;
  sources: string[];
}

export interface IndustryAdoptionRow {
  industry: string;
  year: string;
  adoptionRate: number;
  keyDataPoint: string;
}

export type IndustryAdoptionResponseSummaries = {[key: string]: string};

export interface IndustryAdoptionResponse {
  industries: string[];
  rows: IndustryAdoptionRow[];
  summaries: IndustryAdoptionResponseSummaries;
}

export interface GlobalTrendPoint {
  year: string;
  aiAnyFunctionPct: number;
  genAiSpecificallyPct: number | null;
  topFunction1: string;
}

export interface BusinessFunctionPoint {
  function: string;
  year: string;
  adoptionRate: number;
}

export interface InvestmentPoint {
  country: string;
  year: string;
  investmentB: number;
}

export interface SentimentPoint {
  country: string;
  beneficialPct: number;
  changeJobPct: number | null;
  fearReplacePct: number | null;
}

export interface ToolUserPoint {
  tool: string;
  metricType: string;
  period: string;
  value: number;
  unit: string;
}

export interface ForecastPoint {
  year: number;
  adoptionRate: number;
}

export interface ForecastResponse {
  industry: string;
  historical: ForecastPoint[];
  projected: ForecastPoint[];
  projectedYear: number;
  projectedValue: number;
  gapTo50: number;
  slope: number;
  intercept: number;
  r2: number;
  mae: number;
}

export interface ForecastAllItem {
  industry: string;
  projectedYear: number;
  projectedValue: number;
}

export type GetForecastParams = {
industry: string;
/**
 * @minimum 1
 * @maximum 5
 */
yearsAhead: number;
};

export type GetForecastAllParams = {
/**
 * @minimum 1
 * @maximum 5
 */
yearsAhead: number;
};