# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /forecast:
    get:
      operationId: getForecast
      tags: [forecast]
      summary: Forecast adoption rate for an industry using linear regression
      parameters:
        - name: industry
          in: query
          required: true
          schema:
            type: string
        - name: yearsAhead
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 5
      responses:
        "200":
          description: Forecast result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForecastResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /forecast/all:
    get:
      operationId: getForecastAll
      tags: [forecast]
      summary: Projected adoption rate for every industry at a fixed target year
      parameters:
        - name: yearsAhead
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 5
      responses:
        "200":
          description: Forecast summary per industry
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ForecastAllItem"
        "400":
          description: Invalid industry or yearsAhead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ForecastPoint:
      type: object
      properties:
        year:
          type: number
        adoptionRate:
          type: number
      required: [year, adoptionRate]
    ForecastResponse:
      type: object
      properties:
        industry:
          type: string
        historical:
          type: array
          items:
            $ref: "#/components/schemas/ForecastPoint"
        projected:
          type: array
          items:
            $ref: "#/components/schemas/ForecastPoint"
        projectedYear:
          type: number
        projectedValue:
          type: number
        gapTo50:
          type: number
        slope:
          type: number
        intercept:
          type: number
        r2:
          type: number
        mae:
          type: number
      required:
        - industry
        - historical
        - projected
        - projectedYear
        - projectedValue
        - gapTo50
        - slope
        - intercept
        - r2
        - mae
    ForecastAllItem:
      type: object
      properties:
        industry:
          type: string
        projectedYear:
          type: number
        projectedValue:
          type: number
      required: [industry, projectedYear, projectedValue]