# lib/api-zod/src/generated/types/forecastAllItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ForecastAllItem {
  industry: string;
  projectedYear: number;
  projectedValue: number;
}