# lib/api-zod/src/generated/types/forecastResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ForecastPoint } from './forecastPoint';

export interface ForecastResponse {
  industry: string;
  historical: ForecastPoint[];
  projected: ForecastPoint[];
  projectedYear: number;
  projectedValue: number;
  gapTo50: number;
  slope: number;
  intercept: number;
  r2: number;
  mae: number;
}